backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
openai>=1.40.0
python-dotenv>=1.0.0
